fix: class table modified field

+++ database/migrations/2022_11_24_230652_create_my_classes_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
         Schema::create('my_classes', function (Blueprint $table) {
             $table->id();
             $table->string('name', 100)->unique();
-            $table->unsignedInteger('class_type_id')->nullable()->unique();
+            $table->unsignedBigInteger('class_type_id')->nullable()->unique();
             $table->timestamps();
         });
     }
